requirements.txt
Flask
Jinja2
Flask-SQLAlchemy
Flask-WTF
flask-bcrypt
flask-mysqldb